# Subscriptions search performance fix