@lkraav 2022-06-10
First, improve indexing with https://github.com/OllieJones/index-wp-mysql-for-speed

We were in your exact spot. Better indexes took us from 30s queries to 5s queries, and sometimes even faster.

State of the nation is discussed at https://github.com/WordPress/performance/issues/132

Your strategy improvements proposed might still be valuable even with better indexes.