@AlecKinnear 2022-06-20
Thanks for the link to Ollie's optimisations, Leho. You've always been spot on where we can improve FV Player.

Let's hope we can get performance improvement one through [more smoothly](https://github.com/WordPress/performance/issues/132#issuecomment-1160511202) than the last 30x performance improvement we contributed, which took [ten years](https://core.trac.wordpress.org/ticket/19901#comment:36).