@Jinksi 2022-07-01
@foliovision Thanks for taking the time to raise this issue and open a PR 🙌 

I've opened an issue for this (#183) based on the details you've provided in the PR description. Feel free to make any changes to it based on your experience with the problem.