@haszari 2022-09-04
@AlecKinnear @foliovision can you please add data and detailed reproduce info to  #183? Specifically: 

- How many subscriptions you're running on the site. 
- What plugins and/or customisations you're running, any relevant config.
- The search flow you're using.
- How long it takes for search to complete.
- How long you expect it to take.

Even better, reproduce steps to demonstrate the slow performance on a basic test site with just WooCommerce + WooCommerce Subscriptions (or [WooCommerce Payments with subscriptions active](https://woocommerce.com/document/payments/subscriptions/)). Then we can better understand how this would impact other merchants.

At face value this fix looks reasonable but we need to be clear about the problem we're fixing before we evaluate the fix. 

Thank you!