@AlecKinnear 2022-09-05
@haszari Rue, I'm somewhat astonished at your laundry list of demands here. There's enough demonstrations of the use case above. You have access to dozens of large WooCommerce sites yourself.

What you are asking for sounds to me like a makework project for others to put off doing your own work or integrating the search performance improvements.

I run a business where we build and maintain our own extremely complex plugin, FV Player, an advanced video player for WordPress. I have my own work to do and have never drawn a salary or a cent from Automattic, who has made lots of money on the free work I've already done for them.

If you are in no rush to improve WooCommerce, wish to ignore contributions or stick them into limbo, be my guest.

Neither you nor @m should be surprised about the lack of enthusiasm of other developers to contribute to Automattic projects though.