@haszari 2022-09-05
> There's enough demonstrations of the use case above.

The more information we have about who is affected by this, the better we can prioritise and ensure we're improving performance for all without impacting another area. I figured you'd already have the info about the site where you're experiencing this issue, so it wouldn't be troublesome to provide those details. WooCommerce is a complex system and not all issues affect all sites in the same way, so details are really helpful to us.

If you do have more context or details I'd really appreciate it! If not we will be reviewing the raw PR on its merits.