@ryanmacleod 2022-10-21
Hi, apologies if I'm commenting in the wrong place. 

I'm having the same issue with Search in WooCommerce Subscriptions (approx 30 second response time). 

I see a review has been requested on a potential fix. Does anyone know roughly how long it usually takes before it gets reviewed and ultimately makes it in to a release?