@lkraav 2023-11-22
What are your thoughts on the solution idea provided here @olliejones?